+++ src/Container/HStringLocate.php
@@ -1,4 +1,6 @@
 <?php
+declare(strict_types=0);
+
 namespace Haystack\Container;
 
 use Haystack\HString;
@@ -18,10 +20,14 @@ class HStringLocate
      * @return int
      * @throws ElementNotFoundException
      */
-    public function locate($value)
+    public function locate($value): int
     {
         if ($this->hString->contains($value)) {
-            return mb_strpos($this->hString, (string) $value, 0, $this->hString->getEncoding());
+            $location = mb_strpos($this->hString, (string) $value, 0, $this->hString->getEncoding());
+
+            if ($location !== false) {
+                return $location;
+            }
         }
 
         throw new ElementNotFoundException((string) $value);